Updated more plumes
Kerolox-F1 is in a good spot, as is Kerolox Upper.
Hydrolox SSME isn't needed, would be basically the same as
Hydrolox-Lower.
Hydrolox Lower is reddish, similar in appearance to SSME plume and RS-68
(might need to be a little more pinky-white, but that particular color
particle doesn't exist.
Hydrolox Upper is blue-ish similar in appearance to the J-2X and RL-10
plumes.
Hypergolic-OMS-Grey is redundant, and updates are in work to
Hypergolic-OMS-White.  Will be removing Hypergolic-Apollo-SM, as it'll
basically be the same as the OMS-White.